use new custom markdown renderer
